+++ src/components/menu.js
@@ -1,35 +1,17 @@
-export const createMenuTemplate = (amountFilms) => {
-  return `
-    <nav class="main-navigation">
-      <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-      <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">${amountFilms}</span></a>
-      <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">${amountFilms}</span></a>
-      <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">${amountFilms}</span></a>
-      <a href="#stats" class="main-navigation__item main-navigation__item--additional">Stats</a>
-    </nav>
+import {AbstractComponent} from './abstract-component';
 
-    <ul class="sort">
-      <li><a href="#" class="sort__button sort__button--active">Sort by default</a></li>
-      <li><a href="#" class="sort__button">Sort by date</a></li>
-      <li><a href="#" class="sort__button">Sort by rating</a></li>
-    </ul>
-
-    <section class="films">
-      <section class="films-list">
-        <h2 class="films-list__title visually-hidden">All movies. Upcoming</h2>
-        <div class="films-list__container">
-        </div>
-      </section>
-      <section class="films-list--extra">
-      <h2 class="films-list__title">Top rated</h2>
-        <div class="films-list__container">
-        </div>
-      </section>
-      <section class="films-list--extra">
-      <h2 class="films-list__title">Most commented</h2>
-        <div class="films-list__container">
-        </div>
-      </section>
-    </section>
-  `;
-};
+export class Menu extends AbstractComponent {
+  constructor(amountFilms) {
+    super();
+    this._amountFilms = amountFilms;
+  }
+  getTemplate() {
+    return `<nav class="main-navigation">
+              <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
+              <a href="#watchlist" class="main-navigation__item">Watchlist <span class="main-navigation__item-count">${this._amountFilms}</span></a>
+              <a href="#history" class="main-navigation__item">History <span class="main-navigation__item-count">${this._amountFilms}</span></a>
+              <a href="#favorites" class="main-navigation__item">Favorites <span class="main-navigation__item-count">${this._amountFilms}</span></a>
+              <a href="#stats" class="main-navigation__item main-navigation__item--additional">Stats</a>
+            </nav>>`;
+  }
+}
