+++ src/components/show-more.js
@@ -1,5 +1,11 @@
-export const createShowMoreTemplate = () => {
-  return `
-    <button class="films-list__show-more">Show more</button>
+import {AbstractComponent} from './abstract-component';
+
+export class ShowMore extends AbstractComponent {
+  constructor() {
+    super();
+  }
+  getTemplate() {
+    return `<button class="films-list__show-more">Show more</button>
   `;
-};
+  }
+}
